Fix "opening" state
sometimes "opening" wouldn't stop "opening" when the Valve button was
pressed (see line 270, which had been removed).  Also, if the system was
"locked" (manual mode), and the carriage reached the open end, it would
stop the motor, but not change the state, requiring two V_button
presses.  (see line 305, which was "Open();").  Finally, I re-added
several of the commended-out exit conditions.